@@ -4,16 +4,18 @@ publish_to: "none"
 version: 0.1.0
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  flutter_riverpod: ^2.3.0
+  riverpod: ^2.3.0
 
 dev_dependencies:
+  flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
 
 flutter:
   uses-material-design: true
@@ -45,3 +47,5 @@ flutter:
 
   assets:
     - assets/
+    - assets/images/
+    - assets/images/
